* Fixed: Possible runtime assert
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@9012 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/src/compilersettingsdlg.cpp
@@ -218,7 +218,7 @@ void CompilerSettingsDlg::EndModal(int retCode)
         ConfigManager *bbcfg = Manager::Get()->GetConfigManager(_T("plugins"));
         wxArrayString bbplugins;
         wxCheckListBox* clb = XRCCTRL(*this, "chkBBPlugins", wxCheckListBox);
-        for (int i = 0; i < (int)clb->GetCount(); ++i)
+        for (unsigned int i = 0; i < clb->GetCount(); ++i)
         {
             if (clb->IsChecked(i))
             {
